Move DatasetItem; Remove unnecessary import
