+++ application/ui/src/features/dataset/gallery/media-thumbnail.test.tsx
@@ -3,7 +3,6 @@
 
 import { render, screen, waitFor } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
-import { describe, expect, it } from 'vitest';
 
 import { MediaThumbnail } from './media-thumbnail.component';
