+++ application/ui/src/features/dataset/media-preview/primary-toolbar/zoom/zoom-selector.test.tsx
@@ -2,7 +2,6 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { fireEvent, render, screen } from 'test-utils/render';
-import { describe, expect, it, vi } from 'vitest';
 
 import { useZoom } from '../../../../../components/zoom/zoom.provider';
 import { ZoomSelector } from './zoom-selector.component';
