+++ application/ui/src/features/dataset/media-preview/sidebar-items/sidebar-media-item.component.tsx
@@ -4,8 +4,8 @@
 import { Accept, Search } from '@geti/ui/icons';
 import { useProjectIdentifier } from 'hooks/use-project-identifier.hook';
 import { View } from 'packages/ui';
+import { DatasetItem } from 'src/constants/shared-types';
 import { useAnnotationActions } from 'src/features/annotator/annotation-actions-provider.component';
-import { DatasetItem } from 'src/features/annotator/types';
 
 import { MediaItem } from '../../gallery/media-item.component';
 import { MediaThumbnail } from '../../gallery/media-thumbnail.component';
