+++ application/ui/src/features/dataset/media-preview/sidebar-items/use-keyboard-navigation.hook.test.tsx
@@ -6,7 +6,6 @@ import { useRef } from 'react';
 import { fireEvent } from '@testing-library/react';
 import { getMultipleMockedMediaItems } from 'mocks/mock-media-item';
 import { render, screen } from 'test-utils/render';
-import { vi } from 'vitest';
 
 import { useKeyboardNavigation, UseKeyboardNavigationProps } from './use-keyboard-navigation.hook';
