+++ application/ui/src/features/dataset/toolbar/toolbar.component.tsx
@@ -4,13 +4,13 @@
 import { Button, Divider, Flex, Heading, Text, toast } from '@geti/ui';
 import { useQueryClient } from '@tanstack/react-query';
 import { useProjectIdentifier } from 'hooks/use-project-identifier.hook';
+import { DatasetItem } from 'src/constants/shared-types';
 
 import { $api } from '../../../api/client';
 import { AddMediaButton } from '../../../components/add-media-button/add-media-button.component';
 import { CheckboxInput } from '../../../components/checkbox-input/checkbox-input.component';
-import { useSelectedData } from '../../../routes/dataset/provider';
-import { DatasetItem } from '../../annotator/types';
 import { DeleteMediaItem } from '../gallery/delete-media-item/delete-media-item.component';
+import { useSelectedData } from '../selected-data-provider.component';
 import { toggleMultipleSelection, updateSelectedKeysTo } from './util';
 
 type ToolbarProps = {
