+++ application/ui/src/features/dataset/virtualizer-grid-layout/virtualizer-grid-layout.component.tsx
@@ -7,8 +7,8 @@ import { AriaComponentsListBox, GridLayout, ListBoxItem, Loading, View, Virtuali
 import { useLoadMore } from '@react-aria/utils';
 import { GridLayoutOptions } from 'react-aria-components';
 import { components } from 'src/api/openapi-spec';
-import { MediaState } from 'src/routes/dataset/provider';
 
+import { MediaState } from '../selected-data-provider.component';
 import { useGetTargetPosition } from './use-get-target-position.hook';
 
 import classes from './virtualizer-grid-layout.module.scss';
