+++ application/ui/src/features/inference/sinks/utils.test.ts
@@ -1,8 +1,6 @@
 // Copyright (C) 2025 Intel Corporation
 // SPDX-License-Identifier: Apache-2.0
 
-import { describe, expect, it } from 'vitest';
-
 import { getObjectFromFormData } from './utils';
 
 describe('getObjectFromFormData', () => {
